Maybe this merger is better.
